Решена проблемма с обновлением данных

+++ app/src/main/java/com/example/starwarsapi/feature/domain/usecase/impl/FavoriteUseCaseImpl.kt
@@ -1,6 +1,5 @@
 package com.example.starwarsapi.feature.domain.usecase.impl
 
-import android.util.Log
 import com.example.starwarsapi.feature.domain.model.FavoritePeople
 import com.example.starwarsapi.feature.domain.repository.DatabaseRepository
 import com.example.starwarsapi.feature.domain.usecase.FavoriteUseCase
@@ -10,14 +9,8 @@ class FavoriteUseCaseImpl @Inject constructor(
     private val databaseRepository: DatabaseRepository,
 ) : FavoriteUseCase {
     override suspend fun workDataBase(favoritePeople: FavoritePeople, favorite: Boolean) {
-        if (favorite) {
-            Log.e("Kart","delete")
-            databaseRepository.deleteLike(favoritePeople)
-        }
-        else {
-            Log.e("Kart","set")
-            databaseRepository.setLike(favoritePeople)
-        }
+        if (favorite) databaseRepository.deleteLike(favoritePeople)
+        else databaseRepository.setLike(favoritePeople)
     }
 
 }
